refactor: move reactive_subscriptions from core/ to services/, extract path_utils to lib/
Move ReactiveSubscriptionManager and Subscription to services layer where
they belong (service-level orchestration, not kernel primitives). Extract
path_matches_pattern() to nexus/lib/path_utils.py (analogous to Linux
lib/glob.c) to avoid core→services reverse dependency while keeping it
accessible to all tiers.
