+++ src/nexus/lib/__init__.py
@@ -5,6 +5,7 @@ in ``nexus.contracts`` but are purely implementation helpers rather than
 formal Protocol/ABC contracts.
 
 Modules:
+    path_utils: Cached glob/pattern matching (path_matches_pattern)
     registry: Generic BaseRegistry[T] + BrickRegistry
     rpc_codec: JSON-RPC encode/decode with special-type handling
 """
